refactor(auditor): remove export_test.go anti-pattern and add NewService constructor
Remove export_test.go which was re-exporting unexported symbols solely
for test use — flagged as an anti-pattern by the maintainer.

- Add NewService exported constructor to auditor.go (replaces the
  test-only NewTestService bridge and serves as a proper production
  constructor for dependency injection)
- Move requestWrapper tests into auditor_internal_test.go (package
  auditor) since they access unexported types — a plain move to
  package auditor_test was not possible due to the import cycle:
  auditor/mock imports auditor to implement its interfaces
- External tests in auditor_test.go now use auditor.NewService directly
- Add new tests covering previously uncovered error branches:
  AuditRecord failure, ListAuditTokens error, ToQuantity error,
  RestoreTMS error paths — coverage up to 97.7%
